[OpenScreen] Make Controller manage ConnectRequest
Currently, UrlAvailabilityRequester and MessageGroupStreams have their
own ConnectRequset. It is complicated and not easy to integrate with
authentication in the future. This CL makes Controller manage
ConnectRequest uniformly.

As the result of changes in the process of establishing a connection,
this CL also deletes url_availability_requester_unittest.cc and
moves useful tests to presentation_controller_unittest.cc.

Bug: 281741443
Change-Id: I6c2fea2505c998bd41c0bff85dfdfb08c3a80e92