Fix build and update build dependencies
This patch fixes a critical infra failure where the build is
currently broken due to a minor speed improvement made to the clang
update script.

This patch rolls the build and buildtools submodules as well as updates
the clang update download script to pull a specific version of the
update script, to avoid breaking changes like this in the future.

Change-Id: I975ecaf1667cf03c57ac3f7bd15314b1c54281a9