[Cast] Lower cpu usage by optimizing socket select() call
This patch optimizes a check for Writable handles,
ensuring that we don't pass a handle to the write
argument for select() if we don't have anything to
write. This lowers the amount of cycles spent
polling the select loop.

This is a not-insignificant change, involving updating
the SocketHandleWaiter interface and its subscribers to
allow the waiter to poll for whether the socket has pending
write events or not.

Change-Id: Icbd38a67fbc1fa4856f9605d6306971568dcca5e