Upgrade bots from Mac 13 to Mac 15
Attempting to roll //build currently fails on Mac because
crrev.com/c/6493969 enforced the SDK version to be >= 15 and the bot
only has SDK 14 [1]. In fact the bots are Mac-13 and the mininum mac
version to support SDK 15 is Mac 13.5 [2]. Let's upgrade the bots to
Mac 15 and hope the default SDK version in them is enough to fix the
error. I verified there are 7 Mac 15 machines in the flex pool
(https://crrev.com/i/8754176).

If this is still not enough, I will configure an explicit SDK version
in a follow-up CL.

[1] https://ci.chromium.org/ui/p/openscreen/builders/try/mac_x64/b8698563814246822177/overview
[2] https://developer.apple.com/support/xcode/

Bug: 388070065
Change-Id: If62ddeceba3210f460c8ac91e3961177dedca89d