Update xcode in the bots (2nd attempt)
Update the mac toolchain too. I verified with led that this gets
rid of the permission error, see job link below. That job is using the
pending CL crrev.com/c/7116905 and now shows a different error,
but that can be fixed in the roll CL itself. This CL should be enough to
make mac_arm64 green in CQ.

No-Try because mac_arm64 is broken.

https://ci.chromium.org/ui/p/openscreen/builders/try.shadow/mac_arm64/b8697855589683185745/overview

No-Try: true
Change-Id: Iabf1f48b01d8f35fcf2c85163f95c65d561954bc